requirements.txt
dj-static==0.0.6
Django==1.10.1
django-filter==0.15.0
djangorestframework==3.4.7
mysqlclient==1.3.7
static3==0.7.0
uWSGI==2.0.11.1
